+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - `cached_inject` and `cache_callback` now both accept `float` and `int` seconds for `expire_after`.
-- `OwnerCheck.__init__` now accepts `float` and `int` seconds for `expire_after`.
+- `Owners.__init__` now accepts `float` and `int` seconds for `expire_after`.
+- Rename `tanjun.dependencies.owners.OwnerCheck` and `tanjun.dependencies.owners.AbstractOwnerCheck`
+  to `Owners` and `AbstractOwners`.
 
 ## [2.2.0a1]
 ### Added